+++ tools/autopkg/autopkg_run.go
@@ -1,0 +1,1 @@
+package autopkg
