+++ tools/autopkg/utilities.go
@@ -16,6 +16,32 @@ import (
 	"howett.net/plist"
 )
 
+// Extended environment variables for GitHub Actions integration
+var (
+	USE_JAMF_UPLOADER     bool
+	USE_INTUNE_UPLOADER   bool
+	JAMFPRO_URL           string
+	JAMFPRO_CLIENT_ID     string
+	JAMFPRO_CLIENT_SECRET string
+	INTUNE_TENANT_ID      string
+	INTUNE_CLIENT_ID      string
+	INTUNE_CLIENT_SECRET  string
+	AUTOPKG_REPOS         []string
+	RECIPE_LISTS          []string
+	PRIVATE_REPO_URL      string
+	PRIVATE_REPO_PATH     string
+	JCDS2_MODE            bool
+	API_USERNAME          string
+	API_PASSWORD          string
+	SMB_URL               string
+	SMB_USERNAME          string
+	SMB_PASSWORD          string
+	FORCE_UPDATE          bool
+	FAIL_RECIPES          bool
+	DISABLE_VERIFICATION  bool
+)
+
+// LoadEnvironmentVariables loads all environment variables used by the package
 // LoadEnvironmentVariables loads all environment variables used by the package
 func LoadEnvironmentVariables() {
 	// Check if DEBUG is enabled
@@ -49,6 +75,74 @@ func LoadEnvironmentVariables() {
 	REPORT_PATH = os.Getenv("REPORT_PATH")
 	if REPORT_PATH == "" {
 		REPORT_PATH = "/tmp/autopkg.plist"
+	}
+
+	// GitHub Actions specific variables
+	useJamfStr := os.Getenv("USE_JAMF_UPLOADER")
+	if useJamfStr != "" {
+		USE_JAMF_UPLOADER, _ = strconv.ParseBool(useJamfStr)
+	}
+
+	useIntuneStr := os.Getenv("USE_INTUNE_UPLOADER")
+	if useIntuneStr != "" {
+		USE_INTUNE_UPLOADER, _ = strconv.ParseBool(useIntuneStr)
+	}
+
+	// Jamf Pro settings
+	JAMFPRO_URL = os.Getenv("JAMFPRO_URL")
+	API_USERNAME = os.Getenv("API_USERNAME")
+	API_PASSWORD = os.Getenv("API_PASSWORD")
+	JAMFPRO_CLIENT_ID = os.Getenv("JAMFPRO_CLIENT_ID")
+	JAMFPRO_CLIENT_SECRET = os.Getenv("JAMFPRO_CLIENT_SECRET")
+	SMB_URL = os.Getenv("SMB_URL")
+	SMB_USERNAME = os.Getenv("SMB_USERNAME")
+	SMB_PASSWORD = os.Getenv("SMB_PASSWORD")
+
+	jcds2ModeStr := os.Getenv("JCDS2_MODE")
+	if jcds2ModeStr != "" {
+		JCDS2_MODE, _ = strconv.ParseBool(jcds2ModeStr)
+	}
+
+	// Intune settings
+	INTUNE_TENANT_ID = os.Getenv("INTUNE_TENANT_ID")
+	INTUNE_CLIENT_ID = os.Getenv("INTUNE_CLIENT_ID")
+	INTUNE_CLIENT_SECRET = os.Getenv("INTUNE_CLIENT_SECRET")
+
+	// Repo settings
+	reposStr := os.Getenv("AUTOPKG_REPOS")
+	if reposStr != "" {
+		for _, repo := range strings.Split(reposStr, ",") {
+			repo = strings.TrimSpace(repo)
+			if repo != "" {
+				AUTOPKG_REPOS = append(AUTOPKG_REPOS, repo)
+			}
+		}
+	}
+
+	// Recipe lists
+	listsStr := os.Getenv("RECIPE_LISTS")
+	if listsStr != "" {
+		for _, list := range strings.Split(listsStr, ",") {
+			list = strings.TrimSpace(list)
+			if list != "" {
+				RECIPE_LISTS = append(RECIPE_LISTS, list)
+			}
+		}
+	}
+
+	// Private repo settings
+	PRIVATE_REPO_URL = os.Getenv("PRIVATE_REPO_URL")
+	PRIVATE_REPO_PATH = os.Getenv("PRIVATE_REPO_PATH")
+
+	// General settings
+	forceUpdateStr := os.Getenv("FORCE_UPDATE")
+	if forceUpdateStr != "" {
+		FORCE_UPDATE, _ = strconv.ParseBool(forceUpdateStr)
+	}
+
+	failRecipesStr := os.Getenv("FAIL_RECIPES")
+	if failRecipesStr != "" {
+		FAIL_RECIPES, _ = strconv.ParseBool(failRecipesStr)
 	}
 }
